fix: exhaustive sweep of int vs double mismatches across 10 remaining files

+++ lib/core/utils/excel_export_helper.dart
@@ -106,7 +106,7 @@ class ExcelExportHelper {
   }
 
   /// 3. Official Inventory Audit & Reconciliation Sheet (جدول الجرد الميداني وحساب العجز)
-  static String exportInventoryAuditToCsv(Map<String, int> countedStock) {
+  static String exportInventoryAuditToCsv(Map<String, double> countedStock) {
     final box = HiveDatabase.productBox;
     final products = box.values.toList();
